split out module
