+++ avaframe/ana1Tests/influenceTestCfg.ini
@@ -33,7 +33,6 @@ nPPK = 5
 dt = 0.1
 viscOption = 2
 subgridMixingFactor = 100
-sphOption = 1
 
 [GENERAL]
